Fix broken heartbeat logic
- In the heartbeat tick we check that local time hasn't gone past
the expected heartbeat timeout time. This is very useful on mobile
games where the app can become suspended, as we want to detect as
fast as possible when coming back to the foreground that we've
timed out. Almost all devices are synced to a proper time source,
so this normally works there. On desktops, the local time may vary
wildly from the server, so we need to make sure the timout is
relevant to the local time, taking roundtrip time into account.
- Running with local time wildly off from server time may cause
other problems that we may or may not want to handle, but this at
least fixes the obvious infinite timout, caused by a local clock
that is off by more than the expected heartbeat timeout time.